fix(schema-migrations): fix order of operations, drop views before regular entities [fixes #407]
The RemoveViewDiffer function has been updated to handle both changed and removed views.
It now ensures that views are deleted prior to the entities they may depend on.
This resolves the issue of incorrect operation ordering when both an entity
and a view are deleted within a single migration.

Also, change between regular and view entity is done by the same differ.
