+++ packages/schema-migrations/src/modifications/differs/index.ts
@@ -1,3 +1,2 @@
-export * from './ChangeViewNonViewDiffer'
 export * from './RemoveChangedFieldDiffer'
-export * from './RemoveChangedViewDiffer'
+export * from './RemoveViewDiffer'
